Giao diện sửa thông tin máy tính (chưa cập nhật lại tình trạng linh kiện sau khi thay đổi ở giao diện máy tính)

+++ src/DAL/ComputerDAL.java
@@ -23,7 +23,7 @@ public class ComputerDAL {
                 Computers computer = new Computers(
                         rs.getInt("computer_id"),
                         rs.getString("name"),
-                        rs.getDouble("price_per_hour"),
+                        rs.getInt("price_per_hour"),
                         rs.getInt("motherboard_id"),
                         (Integer)rs.getObject("mouse_id"),
                         (Integer)rs.getObject("keyboard_id"),
